requirements.txt
openai==1.47.1
pymongo==4.9.1
pydantic==2.9.2
python-dotenv~=1.0.1
motor==3.6.0
beanie==1.26.0
fastapi==0.115.0
mongomock_motor==0.0.34